Add some changes Chishm made early on that I missed.
This works for the most part now. Is able to launch games and flashcarts
now. Had to force slot reset though. The ability to card init without
that by not having slot-1 in ntr mode (via the bit for it in access
control) seems broken with this version of libnds. But I can live
without that.

Touchscreen however is borked for NTR games.