Cleaned up some code..
It almost compiles without any warnings now. Only fix left for it that
is to redo slot reset code. Original code is too much of a mess to
compile with the  "-Wall -Wextra -Werror" flag.

That flag has been added to makefile of arm9/bootloader. Once arm7 is
fixed it will be added there as well.