Some fixes
* Small delay added to cart boot process from new hbmenu interface if
user decides to boot a cart that was present on boot and had not been
ejected. (as in that case card init can be skipped which results in very
fast boot time!)
* Card Init now correctly happens as intended if the user booted into
hbmenu interface and decides to boot a card that was present on boot as
in that case the card will need a full reset for boot loader to init it
correctly.